backend/pyproject.toml
[project]
name = "agora"
version = "0.1.0"
description = "Multi-perspective AI council — discuss, design, execute, evolve."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.9.0",
    "sse-starlette>=2.1.0",
    "prompt-toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24.0"]

[project.scripts]
agora = "agora.__main__:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agora"]
